#629 Move DefaultBidirectionalConverter to impl
* Moved DefaultBidirectionalConverter class to impl.
* Updated the import of the unit test.
