+++ platform/dolphin-platform-remoting-client-javafx/src/test/java/com/canoo/dolphin/client/javafx/FXBinderTest.java
@@ -18,9 +18,9 @@ package com.canoo.dolphin.client.javafx;
 import com.canoo.platform.remoting.client.javafx.BidirectionalConverter;
 import com.canoo.platform.core.functional.Binding;
 import com.canoo.platform.remoting.client.javafx.Converter;
-import com.canoo.platform.remoting.client.javafx.binding.DefaultBidirectionalConverter;
 import com.canoo.platform.remoting.client.javafx.FXBinder;
 import com.canoo.platform.remoting.ObservableList;
+import com.canoo.dp.impl.client.javafx.DefaultBidirectionalConverter;
 import com.canoo.dp.impl.remoting.MockedProperty;
 import com.canoo.dp.impl.remoting.collections.ObservableArrayList;
 import com.canoo.platform.remoting.Property;
